add react semantic UI
This changes added the files generated by the semantic-ui and place it in styles directory.

If the `gulp build` executed in the project root directory, it will re-compiles the changes and output it to `vendor` directory

I'm not sure yet if this is the perfect approach to manage the semantic-ui related files, but let's keep it that way for now while searching for the best options that fit to our needs down the road.