update the retrieveServicePictures action
this action will get service picture for each of the service object cached on the state.
after, the images has been added to each of the service object, then dispatch an action to update the popular places,
to make sure the the popular places contains the place services with an image attached to it

Do not use catch, because that will also catch any errors in the dispatch and resulting render,
causing a loop of 'Unexpected batch number' errors.
https://github.com/facebook/react/issues/6895